Feat/model batch services (#62)
* test(DAL): rename model service base tests class

* test(DAL): reorganize class methods in base class ModelServiceTest

* test(DAL): fix minor formatting

* test(DAL): add base class for batch model service tests

* test(DAL): remove unnecessary property

* test(DAL): add documentation to test methods

* fix: change BrokerTransactionCode attribute to allow negative

Changes effect attributes of BrokerTransactionCode model from byte to short.
Byte data type is correctly mapped to tinyint in SQL Server, but does not allow
negative values.

* test(DAL): add test/reference data models for bank/broker codes

* build: change test project name and namespace

* test(DAL): remove test data from production code

* refactor: revise seed data implementation

* test(DAL): move test models into seed data implementation

* test(DAL): add clean-up method to reset test db to default state

* build: rename file directories

* refactor: add null/empty handle to model seed method

Also move ISeedData to more appropriate namespace.

* refactor: remove obsolete abstract methods

* fix: add AccountWallet collection to seed data

* test(DAL): add AccountWalletServiceTest

* refactor(DAL): move DbContext property to base class

* fix: add support for shared context among reader/writer model services

Changes model services to allow for sharing context between reader and
writer services, simplifying change tracking in batch models services.

* fix(DAL): fix property accessor for ModelWriterBatchService

* docs(DAL): fix in-code documentation for IModelBatchService

* feat(DAL): add child batch service configuration extension method

Adds extension method to IModelBatchService interface allowing for fluid
configuration of the parent id.

* test(DAL): refactor model service tests and add new batch service test

* fix(DAL): add remaining model collections to ISeedData

Updates ModelServiceTestDataModel to implement changes to
ISeedData.

* fix(DAL): add ModelBuilder support for remaining ISeedData members

* fix(DAL): fix error in the ModelBuilder data-seeding method

* test(DAL): move model service test configuration namespace

* feat(DAL): add AccountAttributMemberService

Resolves #36

* test(DAL): add AccountAttributeMemberServiceTest

Related #36

* test(DAL): change to compound assignment for test variable

* feat(DAL): add AccountCompositeMemberService

Resolves #35

* feat: add BankTransactionCodeAttributeService

Resolves #33

* feat(DAL): add BankTransactionService

Resolves #34

* feat(DAL): add BrokerTransactionCodeAttributeService

Resolves #31

* feat(DAL): add BrokerTransactionService

Resolves #32

* feat(DAL): add CountryAttributeService

Resolves #30

+++ NjordFinance.DAL/Model/BrokerTransactionCode.cs
@@ -24,9 +24,9 @@ namespace NjordFinance.Model
         [Required]
         [StringLength(32)]
         public string DisplayName { get; set; }
-        public byte CashEffect { get; set; }
-        public byte ContributionWithdrawalEffect { get; set; }
-        public byte QuantityEffect { get; set; }
+        public short CashEffect { get; set; }
+        public short ContributionWithdrawalEffect { get; set; }
+        public short QuantityEffect { get; set; }
 
         [InverseProperty(nameof(BrokerTransactionCodeAttributeMemberEntry.TransactionCode))]
         public virtual ICollection<BrokerTransactionCodeAttributeMemberEntry> BrokerTransactionCodeAttributeMemberEntries { get; set; }
